+++ core/core-domain/build.gradle
@@ -1,15 +1,15 @@
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
-    // Redis
-    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-redis', version: '3.2.0'
-    implementation 'org.springframework.session:spring-session-data-redis:3.1.1'
-
     //querydsl
     implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
     annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
     annotationProcessor "jakarta.annotation:jakarta.annotation-api"
     annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+
+    // Redis
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-redis', version: '3.2.0'
+    implementation 'org.springframework.session:spring-session-data-redis:3.1.1'
 }
 
 bootJar { enabled = false }
